+++ csatar-plugins/csatar/lang/en/lang.php
@@ -194,6 +194,17 @@
                 'foodSensitivities' => 'Food Sensitivities',
                 'breadcrumb' => 'Food Sensitivities',
             ],
+            'contactSettings' => [
+                'contactSettings' => 'Contact Settings',
+                'description' => 'Data from the Contact Us page.',
+                'offices' => 'Offices',
+                'promptNew' => 'Add new item',
+                'address' => 'Address',
+                'bank' => 'Bank',
+                'bankAccount' => 'Bank Account',
+                'email' => 'Email',
+                'phoneNumbers' => 'Phone Numbers',
+            ],
             'chronicIllness' => [
                 'chronicIllness' => 'Chronic Illness',
                 'chronicIllnesses' => 'Chronic Illnesses',
@@ -423,7 +434,8 @@
             ],
             'permissions' => [
                 'permissions' => 'Permissions',
-                'allPermissionsForScout' => 'All permissions for every model and field for Scout mandate.'
+                'allPermissionsForScout' => 'All permissions for every model and field for Scout mandate.',
+                'readPermissionForGuests' => 'Read permission for guests.',
             ],
         ],
         'component' => [