+++ csatar-plugins/csatar/lang/hu/lang.php
@@ -194,6 +194,17 @@
                 'foodSensitivities' => 'Ételérzékenységek',
                 'breadcrumb' => 'Ételérzékenységek',
             ],
+            'contactSettings' => [
+                'contactSettings' => 'Kapcsolat beállítások',
+                'description' => 'Adatok a Kapcsolat oldalról.',
+                'offices' => 'Irodák',
+                'promptNew' => 'Új hozzáadása',
+                'address' => 'Cím',
+                'bank' => 'Bank',
+                'bankAccount' => 'Bankszámla',
+                'email' => 'E-mail cím',
+                'phoneNumbers' => 'Telefonszámok',
+            ],
             'chronicIllness' => [
                 'chronicIllness' => 'Krónikus betegség',
                 'chronicIllnesses' => 'Krónikus betegségek',
@@ -423,7 +434,8 @@
             ],
             'permissions' => [
                 'permissions' => 'Jogosultságok',
-                'allPermissionsForScout' => 'Minden jogosultság a Cserkész megbízáshoz, az összes model összes mezőjéhez.'
+                'allPermissionsForScout' => 'Minden jogosultság a Cserkész megbízáshoz, az összes model összes mezőjéhez.',
+                'readPermissionForGuests' => 'Olvasási jog vendégeknek.',
             ],
         ],
         'component' => [