+++ csatar-plugins/csatar/models/MandateType.php
@@ -8,6 +8,8 @@ use Csatar\Csatar\Models\Troop;
 use Input;
 use Lang;
 use Model;
+use October\Rain\Database\Collection;
+use Session;
 
 /**
  * Model
@@ -143,21 +145,43 @@ class MandateType extends Model
         return $association_id && $mandate_model_type ? $query->where('association_id', $association_id)->where('organization_type_model_name', $mandate_model_type) : $query->whereNull('id');
     }
 
-    function scopeMandateTypeIdsInAssociation($query, $associationId) {
+    public static function getAllMandateTypeIdsInAssociation($associationId)
+    {
         return self::where('association_id', $associationId)->get()->pluck('id');
     }
 
-    function scopeScoutMandateTypeIdInAssociation($query, $associationId): array {
+    public static function getScoutMandateTypeIdInAssociation($associationId): array
+    {
         $scoutMandateType = self::where('association_id', $associationId)
             ->where('organization_type_model_name', '\Csatar\Csatar\Models\Scout')
             ->first();
         return $scoutMandateType ? [ $scoutMandateType->id ] : [];
     }
 
-    function scopeGuestMandateTypeInAssociation($query, $associationId): array {
-        $guestMandateType = self::where('association_id', $associationId)
+    public static function getGuestMandateTypeIdInAssociation($associationId): ?int
+    {
+        $sessionRecord = Session::get('guest.mandateTypeIds');
+
+        if(!empty($sessionRecord) && $sessionRecordForAssociation = $sessionRecord->where('associationId', $associationId)->first()) {
+            return $sessionRecordForAssociation['guestMandateTypeId'];
+        }
+
+        if(empty($sessionRecord)){
+            $sessionRecord = new Collection([]);
+        }
+
+        $guestMandateTypeId = self::where('association_id', $associationId)
             ->where('organization_type_model_name', self::MODEL_NAME_GUEST)
-            ->first();
-        return $guestMandateType ? [ $guestMandateType->id ] : [];
+            ->first()->id;
+
+        $sessionRecord = $sessionRecord->replace([ $associationId => [
+            'associationId' => $associationId,
+            'savedToSession' => date('Y-m-d H:i'),
+            'guestMandateTypeId'=> $guestMandateTypeId,
+        ]]);
+
+        Session::put('guest.mandateTypeIds', $sessionRecord);
+
+        return $guestMandateTypeId;
     }
 }