+++ csatar-theme/assets/csat-jquery/csat-data-card.js
@@ -1,0 +1,5 @@
+$( document ).ready(function() {
+	if ($(document).width()>767){
+		$(".collapse-on-sm").removeClass("collapse"); 
+	} 
+});